// src/generated/types.ts
export interface AddCirTicketRequest {
  cirInvoiceId?: string | null
  data?: string | null
  amount?: number | null
  userComment?: string | null
  resourceType: CirResourceType
  cirTicketCategory: CirTicketCategory
}
// ...
export interface ChangeUser {
  firstName?: string | null
  lastName?: string | null
}

export interface CirHistoryDto {
  comment?: string | null
  assignment: AssignmentDto
  amountChanges?: AmountChangeDto[] | null
  cancellation: CancellationDto
  settlements?: SettlementDto[] | null
}

export interface CirTicketChangeDto {
  id: number
  propertyName?: string | null
  oldValue?: string | null
  newValue?: string | null
  dateChanged: string
  user: ChangeUser
  version: number
  serviceDesk: boolean
}

export interface CirTicketDto {
  id: number
  cirId?: string | null
  category?: string | null
  data?: string | null
  organizationId: number
  userComment?: string | null
  operatorComment?: unknown | null
  status: CirTicketStatus
  creationDate: string
  closingDate?: string | null
  categoryCyr?: string | null
  companyNumber?: string | null
  organizationName?: string | null
  resourceId: CirResourceType
  cirTicketHistory: CirTicketHistoryDto
}

export interface CirTicketHistoryDto {
  cirTicketId: number
  cirTicketChanges?: CirTicketChangeDto[] | null
}

export interface CirTicketListResponse {
  total: number
  tickets?: CirTicketDto[] | null
}

export interface CirTicketSearchEnumRestrictionItem {
  field: CirTicketSearchEnum
  values?: string[] | null
}

export interface CirTicketSearchParameter {
  sortItems?: CirTicketSortEnumSortItem[] | null
  pagingOptions: PagingOptions
  restrictions?: CirTicketSearchEnumRestrictionItem[] | null
}

export interface CirTicketSortEnumSortItem {
  sortColumn: CirTicketSortEnum
  sortDirection: SortDirection
}

export interface CompanyAccountIdentificationDto {
  registrationNumber?: string | null
  jbkjs?: string | null
  vatNumber?: string | null
}

export interface CompanyAccountOnEfAkturaDto {
  eFakturaRegisteredCompany: boolean
}

export interface CompanyDetailsRequest {
  vatRegistrationCode?: string | null
  registrationCode?: string | null
}
// ...
export interface CompanyFullDetailsDto {
  id: number
  vatRegistrationCode?: string | null
  name?: string | null
  address?: string | null
  registrationCode?: string | null
  isBudget: boolean
}
// ...
export interface CustomsDeclarationDto {
  id: number
  year: number
  transitOperationRegistrationNumber: number
  transitOperationOfficeCode?: string | null
  documentCode?: string | null
  version: number
  taxIdReceiver?: string | null
  taxId54?: string | null
  originCountryCode?: string | null
  declarationAcceptanceDate: string
  declarationSubmissionDate: string
  customsDate: string
  sysDate: string
  itemCount: number
  status: CustomsDeclarationStatus
  duties: CustomsDeclarationPaymentStatus
  taxTotal?: number | null
  dutyTotal?: number | null
  versionsAvailable?: number[] | null
  customsInvoiceNumber?: string | null
  originCountriesCodes?: string[] | null
  summedDtos?: CustomsDeclarationDutySummedDto[] | null
  minFinCodeValuesSummedDtos?: CustomsDeclarationMinFinCodeValuesSummedDto[] | null
  paymentsResponse: CustomsDeclarationPaymentsDto
}

export interface CustomsDeclarationDutySummedDto {
  vatCategoryCode?: string | null
  dutyValueSummedType: DutyValueSummedType
  baseValue?: number | null
  value?: number | null
  rate?: number | null
}

export interface CustomsDeclarationIdsResponse {
  customsDeclarationIds?: number[] | null
}

export interface CustomsDeclarationItemDocumentDto {
  documentCode?: string | null
  currencyCode?: string | null
  ordinalNumber: number
  documentValue: number
  additionalInfo?: string | null
}

export interface CustomsDeclarationItemDto {
  value: number
  exNum?: string | null
  ordinalNumber: number
  bindingTariffReferenceValue: number
  taxExemptionCode?: string | null
  taxExemptionMinFinCode?: string | null
  originCountryCode?: string | null
  taxExemptionReasonCode?: string | null
  sysDate?: string | null
  customsDeclarationItemDocuments?: CustomsDeclarationItemDocumentDto[] | null
  customsDeclarationItemDuties?: CustomsDeclarationItemDutyDto[] | null
}

export interface CustomsDeclarationItemDutyDto {
  dutyCode?: string | null
  ordinalNumber: number
  baseValue: number
  rate: number
  dutyValue: number
}

export interface CustomsDeclarationMinFinCodeValuesSummedDto {
  minFinCode?: string | null
  baseValue?: number | null
}

export interface CustomsDeclarationPaymentDto {
  year?: number | null
  transitOperationRegistrationNumber?: number | null
  transitOperationOfficeCode?: string | null
  documentCode?: string | null
  version?: number | null
  taxPayerTaxId?: string | null
  taxDutyPayment: number
  customsDutyPayment: number
  exciseDutyPayment: number
  otherDutyPayment: number
  totalAmountPaid: number
  paymentDate: string
  customsDeclarationStatus: CustomsDeclarationStatus
}

export interface CustomsDeclarationPaymentsDto {
  totalPaidAmount?: number | null
  totalVatPaidAmount?: number | null
  totalAmountToPay?: number | null
  balance?: number | null
  remainingVatToPay?: number | null
  remainingExcisesToPay?: number | null
  remainingCustomsToPay?: number | null
  remainingOtherDutiesToPay?: number | null
  payments?: CustomsDeclarationPaymentDto[] | null
}

export interface CustomsDeclarationStatusChangeResponse {
  transitOperationRegistrationNumber: number
  transitOperationOfficeCode?: string | null
  year: number
  documentType?: string | null
  version: number
  changeDate: string
  oldStatus: CustomsDeclarationPaymentStatus
  newStatus: CustomsDeclarationPaymentStatus
}
// ...
export interface FiscalBillItemOutputApiDto {
  quantity: number
  total: number
  labels?: string | null
  unitPrice: number
  name?: string | null
}

export interface FiscalBillOutputApiDto {
  tin: string
  buyer?: string | null
  buyerAdditionalField?: string | null
  transactionType: number
  sdcTimestamp?: string | null
  referentDocumentNumber?: string | null
  referentDocumentSdcTimestamp?: string | null
  taxRateGroupId?: number | null
  totalAmount: number
  fiscalBillNumber?: string | null
  status: number
  payments?: FiscalBillPaymentOutputApiDto[] | null
  invoiceItems?: FiscalBillItemOutputApiDto[] | null
  invoiceTaxRateItems?: FiscalBillTaxRateItemOutputApiDto[] | null
}

export interface FiscalBillPaymentOutputApiDto {
  amount: number
}

export interface FiscalBillSimpleDto {
  fiscalBillId: number
  fiscalBillNumber?: string | null
  sellerTin?: string | null
  sellerName?: string | null
  buyerTin?: string | null
  buyerName?: string | null
  transactionType: number
  issueDate?: string | null
  totalAmount: number
  isInvoiceCreated: number
  invoiceNumber?: string | null
  invoiceSentDate?: string | null
}

export interface FiscalBillTaxRateItemOutputApiDto {
  rate: number
  label?: string | null
}

export interface FiscalBillViewDto {
  fiscalBills?: FiscalBillSimpleDto[] | null
}

export interface FullReceiverRecipientsNoticeAPIDto {
  noticeNumber?: string | null
  noticeDate?: string | null
  totalVatAmount?: number | null
  documentIssueOrigin: DocumentIssueOrigin
  basisOfNotice: BasisOfNotice
  documentReferenceType: DocumentReferenceTypes
  relatedDocumentNumber?: string | null
  relatedDocumentIssueDate?: string | null
  relatedInvoicePeriodStartDate?: string | null
  relatedInvoicePeriodEndDate?: string | null
  relatedDocumentStornoDate?: string | null
  notes?: string | null
  id?: number | null
  sender: CompanyFullDetailsDto
  recipient: CompanyFullDetailsDto
  salesInvoiceId?: number | null
  purchaseInvoiceId?: number | null
}

export interface FullSenderRecipientsNoticeAPIDto {
  noticeNumber?: string | null
  noticeDate?: string | null
  totalVatAmount?: number | null
  documentIssueOrigin: DocumentIssueOrigin
  basisOfNotice: BasisOfNotice
  documentReferenceType: DocumentReferenceTypes
  relatedDocumentNumber?: string | null
  relatedDocumentIssueDate?: string | null
  relatedInvoicePeriodStartDate?: string | null
  relatedInvoicePeriodEndDate?: string | null
  relatedDocumentStornoDate?: string | null
  notes?: string | null
  id?: number | null
  recipient: CompanyFullDetailsDto
  sender: CompanyFullDetailsDto
  sendingStatus: SenderRecipientsNoticeStatus
  salesInvoiceId?: number | null
  purchaseInvoiceId?: number | null
}
// ...
export interface PagingOptions {
  pageIndex: number
  pageSize: number
}
// ...
export interface PurchaseInvoicesDto {
  purchaseInvoiceIds?: number[] | null
}

export interface PurchaseInvoiceStatusChangeDto {
  eventId: number
  date?: string | null
  newInvoiceStatus: PurchaseInvoiceStatus
  purchaseInvoiceId: number
  comment?: string | null
  cirInvoiceId?: string | null
  subscriptionKey?: string | null
  stornoNumber?: string | null
  cirAssignmentChange: CirAssignmentChange
  isSigned?: boolean | null
  isAutoAssigned: boolean
}
// ...
export interface SenderRecipientsNoticeRequest {
  noticeNumber?: string | null
  totalVatAmount?: number | null
  documentIssueOrigin: DocumentIssueOrigin
  basisOfNotice: BasisOfNotice
  documentReferenceType: DocumentReferenceTypes
  relatedDocumentNumber?: string | null
  relatedDocumentIssueDate?: string | null
  relatedInvoicePeriodStartDate?: string | null
  relatedInvoicePeriodEndDate?: string | null
  relatedDocumentStornoDate?: string | null
  notes?: string | null
  recipient: CompanyDetailsRequest
}

export interface SettlementDto {
  settlementDate?: string | null
  amount: number
  comment?: string | null
}

export interface SimplePurchaseInvoiceDto {
  invoiceId: number
  globUniqId?: string | null
  comment?: string | null
  cirStatus: CirInvoiceStatus
  cirInvoiceId?: string | null
  version: number
  lastModifiedUtc: string
  cirSettledAmount: number
  vatNumberFactoringCompany?: string | null
  factoringContractNumber?: string | null
  cancelComment?: string | null
  stornoComment?: string | null
  status: PurchaseInvoiceStatus
}